added position control submodes
in position control mode the user can now choose between two different
submodes. The first mode is standard position control. The second mode is
circle mode in which the vehicle flies a circle around a given point. The
user can change the shape of the circle and command the velocity with which
the drone will fly along the circle.